Time: 94 ms (67.37%) | Memory: 99.1 MB (16.98%) - LeetSync

+++ 3332-minimum-operations-to-exceed-threshold-value-ii/minimum-operations-to-exceed-threshold-value-ii.cpp
@@ -17,6 +17,6 @@ public:
     cnt++;
   }
 
-  return cnt;
+  return (int)cnt;
     }
 };
